fix, debug, stores update, better architecture, config done
