+++ front-end/src/lib/store.js
@@ -7,5 +7,3 @@ export const email = writable('');
 export const authStage = writable(false);
 export const authSuccess = writable(false);
 export const networth = writable([]);
-export const networthProgressLineChartArray = writable([])
-export const networthProgressLineChartDatesArray = writable([])
