+++ front-end/src/routes/_auth/+page.svelte
