+++ front-end/src/routes/_submit/+page.svelte
@@ -1,0 +1,7 @@
+<div class="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50 overflow-hidden">
+  <!-- Popup Content -->
+  <div class="bg-white p-4 shadow-md rounded-md text-black">
+    <!-- Your Popup Content Goes Here -->
+    <p>This is your popup content.</p>
+  </div>
+</div>
