+++ front-end/tailwind.config.js
@@ -2,7 +2,7 @@ const colors = require('tailwindcss/colors')
 
 /** @type {import('tailwindcss').Config} */
 export default {
-  content: ['./src/**/*.{html,js,svelte,ts}'],
+  content: ['./src/**/**/*.{html,js,jsx,ts,tsx,svelte}'],
   theme: {
     extend: {
       colors: {
